test: Fix unit tests
