--- integration_test/utils/expectations.dart
@@ -1,23 +1,0 @@
-import 'package:flutter_test/flutter_test.dart';
-import 'package:gymtracker/model/exercise.dart';
-
-void expectAbstractExercise(Exercise result, Exercise expected) {
-  expect(result.id, expected.id);
-  expect(result.name, expected.name);
-  expect(result.parameters, expected.parameters);
-  expect(result.primaryMuscleGroup, expected.primaryMuscleGroup);
-  expect(result.secondaryMuscleGroups, expected.secondaryMuscleGroups);
-}
-
-void expectExercise(Exercise result, Exercise expected) {
-  expectAbstractExercise(result, expected);
-  expect(result.restTime, expected.restTime);
-  expect(result.notes, expected.notes);
-  expect(result.sets.length, expected.sets.length);
-  for (int i = 0; i < result.sets.length; i++) {
-    expect(result.sets[i].kind, expected.sets[i].kind);
-    expect(result.sets[i].parameters, expected.sets[i].parameters);
-    expect(result.sets[i].reps, expected.sets[i].reps);
-    expect(result.sets[i].weight, expected.sets[i].weight);
-  }
-}
